pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "James Fitzgerald", email = "james@jawfish.dev" },
    { name = "Shaikha Aburehaima", email = "shaikha.ab0970@gmail.com" },
]
dependencies = [
    "nltk>=3.9.1",
    "openai>=1.51.0",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.12"


[tool.rye.scripts]
app = "python src/chatbot/main.py"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["ruff>=0.6.8"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/chatbot"]
